Re-implement paperdoll/context menu. Add support for Paperdoll/Whisper/Friend/Ignore context menu actions. Support unequipping items from paperdoll.

+++ EOBot/TrainerBot.cs
@@ -343,7 +343,7 @@ namespace EOBot
 
             ConsoleHelper.WriteMessage(ConsoleHelper.Type.UseItem, $"{itemToUse.Name} - {itemToUse.HP} HP - inventory: {amount - 1} - (other heal item types: {healItems.Count() - 1})");
 
-            await TrySend(() => _characterActions.UseItem(itemToUse.ID));
+            await TrySend(() => _characterActions.UseItem((short)itemToUse.ID));
 
             await Task.Delay(ATTACK_BACKOFF_MS);
         }
